+++ src/App.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import Login from './components/Login';
+
+const main = document.getElementById('main');
+ReactDOM.render(<Login />, main);