+++ webpack.config.js
@@ -1,8 +1,8 @@
 module.exports = {
-  entry: './src/main.js',
+  entry: './src/App.js',
   output: {
     filename: 'dist.js', // 打包文件名
-    path: `${__dirname}/build`, // webpack本地打包路径,与publicPath作用不同
+    path: `${__dirname}/public/dist`, // webpack本地打包路径,与publicPath作用不同
     // chunkFilename: '[name]-[id]-chunk.js',
     // 运行服务器时的打包文件夹路径,即打包在 "http://网站根目录/dist/"下,通过"http://网站根目录/dist/bundle.js"访问.
     publicPath: '/dist/', // http://www.jb51.net/article/116443.htm  publicPath路径问题详解
@@ -14,7 +14,7 @@ module.exports = {
     // historyApiFallback: true, // false 根据路径跳转页面 ; true 路径都返回根index.html
     port: 8088,
     open: true,
-    hot: true,
+    // hot: true,
     watchOptions: {
       aggregateTimeout: 1000,
       ignored: /node_modules/,