+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Version 1.0.0 (2018-03-12)
+# Version 1.0.0 (2018-03-14)
 
 
 ### chore
@@ -8,6 +8,15 @@
 ### code
 
 * source code ([af0dcb8](https://github.com/Xiongqi-XQ/Draven-System/commit/af0dcb8)) (XiongQi)
+
+### feat
+
+* project init ([e9f64e2](https://github.com/Xiongqi-XQ/Draven-System/commit/e9f64e2)) (XiongQi)
+
+### test
+
+* login test ([202f660](https://github.com/Xiongqi-XQ/Draven-System/commit/202f660)) (XiongQi)
+* mobx-react test ([d33da2f](https://github.com/Xiongqi-XQ/Draven-System/commit/d33da2f)) (XiongQi)